+++ lib/type_class/class/protocol.ex
@@ -1,10 +1,10 @@
-defmodule Superclass.Class.Protocol do
+defmodule TypeClass.Class.Protocol do
   @moduledoc ~S"""
   The protocol helpers for defining the critical functions of a type class
   """
 
-  alias Superclass.Utility
-  use Superclass.Utility.Attribute
+  alias TypeClass.Utility
+  use TypeClass.Utility.Attribute
 
   defmacro __using__(_) do
     quote do
