+++ test/type_class_test.exs
@@ -1,7 +1,7 @@
-defmodule SuperclassTest do
+defmodule TypeClassTest do
   use ExUnit.Case
 
-  doctest Superclass
+  doctest TypeClass
 
   test "the truth" do
     assert 1 + 1 == 2
